requirements.txt
fastapi==0.103.2
GeoAlchemy2==0.14.1
keycloak==3.1.3
pydantic==2.4.2
python-dotenv==1.0.0
python_jose==3.3.0
Requests==2.31.0
SQLAlchemy==2.0.21
